Swiched from database_tables.js to query_metadata.js
