+++ lib/querytables.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const DatabaseTables = require('./models/database_tables');
+const QueryMetadataModel = require('./models/query_metadata');
 const SubstitutionTokens = require('./utils/substitution_tokens');
 
 function replaceTile0 (sql) {
@@ -25,9 +25,9 @@ module.exports.getQueryMetadataModel = function (pg, sql, callback) {
         result = result || {};
         const rows = result.rows || [];
 
-        callback(null, new DatabaseTables(rows));
+        callback(null, new QueryMetadataModel(rows));
     });
 };
 
 
-module.exports.DatabaseTablesEntry = DatabaseTables;
+module.exports.QueryMetadata = QueryMetadataModel;
